Update to work with puppet-lint >= 2.4
puppet-lint 2.4 introduced what was supposed to be a non-breaking change
that fixed how interpolated expressions were tokenised, but this
unfortunately broke the check in this plugin (I'm going to have to
rethink how plugins factor into my change classifications for semver
purposes).

Anyway, this PR should fix up the issues with this plugin and
puppet-lint 2.4.x by extending the existing logic to handle cases where
variables like `$facts['osfamily']` are represented internally as
a sequence of tokens rather than a single token.